Fixed incorrect addition of equipment to the hero build

+++ app/src/main/java/com/example/tanorami/equipment/data/EquipmentRepositoryImpl.kt
@@ -17,7 +17,7 @@ class EquipmentRepositoryImpl @Inject constructor(
 ) : EquipmentRepository {
     override suspend fun getWeapons(path: Int): List<Equipment> = withContext(ioDispatcher) {
         weaponDao.getWeaponByPath(path).sortedByDescending { it.rarity }.map {
-            Equipment(it.idWeapon, it.image, it.rarity.toByte())
+            Equipment(it.idWeapon, it.image, it.rarity)
         }
     }
