+++ app/src/main/java/com/example/tanorami/equipment/data/model/Equipment.kt
@@ -4,4 +4,4 @@ import android.os.Parcelable
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
-data class Equipment(val id: Int, val image: String = "", val rarity: Byte = 2) : Parcelable
+data class Equipment(val id: Int, val image: String = "", val rarity: Int = 2) : Parcelable
